refactor: start typing all the Strings (#247)
This is sort of a proposal. Following the comment in
https://github.com/coursier/jvm-index/pull/243#issuecomment-2007605083
by @carlosedp I was going to try and write some docs on the actual
structure of the index, but right away hit on the fact that it's just
this:

```
final case class Index(map: Map[String, Map[String, Map[String, Map[String, String]]]]) {
```

It's pretty impossible to know what these `String`s all are so I figured
maybe it'd be good to start just typing this out a bit more. I used an
opaque type for `Os` here in hopes that it'd make things clearer a bit
throughout the code, and figured we could do that for all the `String`s
in the `Index`, but wanted to see what it would be like to at least do
one of them first. Thoughts on this?